feat: Add Hero component and update metadata in layout
The Hero component is added to the Hero.tsx file, which renders a hero section with multiple Spotlight components. The Spotlight component is also added to the ui folder. This component will be used to showcase important content on the website.

The metadata in the layout.tsx file is updated to reflect the new branding and description of the website.

chore: Update tailwind colors configuration and npm dependencies

The tailwind.config.ts file is updated to include a new color variant for the black color, providing more flexibility in using it in different contexts. Additionally, the package.json file is updated to include new dependencies clsx and tailwind-merge, required for the newly added utility function and for merging Tailwind CSS classes.

chore: Remove unused CSS styles

The globals.css file is updated to remove unused CSS styles related to background colors and text wrapping, improving performance.

chore: Update tailwind colors configuration

The tailwind colors configuration is updated to include a new shade for the black color.

chore: Update npm dependencies

The package.json file is updated to include new dependencies clsx and tailwind-merge.

base file and folder structure

rm

add